Polar: Win32 code now updates with set frame rate

+++ project/code/polar_buffer.cpp
@@ -1,6 +1,7 @@
 #ifndef polar_buffer_cpp
 #define polar_buffer_cpp
 
+
 POLAR_RINGBUFFER *polar_ringbuffer_Create(MEMORY_ARENA *Arena, u32 Samples, u32 Blocks)
 {
     if(Blocks <= 0)
@@ -13,7 +14,7 @@ POLAR_RINGBUFFER *polar_ringbuffer_Create(MEMORY_ARENA *Arena, u32 Samples, u32 
     
     Result->Samples = Samples;
     Result->TotalBlocks = Blocks;
-    Result->Data = (f32 *) memory_arena_Push(Arena, Result->Data, ((sizeof (*Result->Data)) * (Result->Samples * Result->TotalBlocks)));
+    Result->Data = (i16 *) memory_arena_Push(Arena, Result->Data, ((sizeof (*Result->Data)) * (Result->Samples * Result->TotalBlocks)));
     
     Result->CurrentBlocks = 0;
     Result->ReadAddress = 0;
@@ -37,7 +38,7 @@ void polar_ringbuffer_Destroy(MEMORY_ARENA *Arena, POLAR_RINGBUFFER *Buffer)
     }
 }
 
-f32 *polar_ringbuffer_WriteData(POLAR_RINGBUFFER *Buffer)
+i16 *polar_ringbuffer_WriteData(POLAR_RINGBUFFER *Buffer)
 {
     return Buffer->Data + (Buffer->WriteAddress * Buffer->Samples);
 }
@@ -53,7 +54,7 @@ void polar_ringbuffer_WriteFinish(POLAR_RINGBUFFER *Buffer)
     Buffer->CurrentBlocks += 1;
 }
 
-f32 *polar_ringbuffer_ReadData(POLAR_RINGBUFFER *Buffer)
+i16 *polar_ringbuffer_ReadData(POLAR_RINGBUFFER *Buffer)
 {
     return Buffer->Data + (Buffer->ReadAddress * Buffer->Samples);
 }
